create a new file to save the email extension
